airflow/requirements.txt
apache-airflow==2.9.0
pandas
requests
psycopg2-binary